+++ packages/board/src/hooks/index.ts
@@ -1,4 +1,6 @@
 export { useKanbanState } from './useKanbanState'
+export { useBoard } from './useBoard'
+export { useFilters } from './useFilters'
 export { useAI } from './useAI'
 export { useMultiSelect } from './useMultiSelect'
 export {
@@ -12,6 +14,8 @@ export {
 export { useKeyboardShortcuts, DEFAULT_SHORTCUTS } from './keyboard/useKeyboardShortcuts'
 
 export type { UseKanbanStateOptions, UseKanbanStateReturn } from './useKanbanState'
+export type { UseBoardOptions, UseBoardReturn } from './useBoard'
+export type { UseFiltersOptions, UseFiltersReturn, FilterState, SortState, DateFilter, SortBy, SortOrder } from './useFilters'
 export type { UseAIOptions, UseAIReturn } from './useAI'
 export type { UseMultiSelectReturn } from './useMultiSelect'
 export type {